Add more logging and pass project from trigger
Add more logging when build fails because we can't find the correct SCM
configuration. Also pass the project from the trigger into DescriptorImpl
instead of using whatever is in there. This seems to resolve a lot of issues
where DescriptorImpl has the incorrect project and is thrown off by non
git projects in jenkins.