Upgrade to use GitLab APIv4
This commit is the simplest change which seems to suffice to move the
Jenkins plugin to APIv4. [The docs][api-docs] regarding the move from
v3 to v4 and going through most of the source, suggest that is the only
change too.

To summerize, the path changed from `api/v3` to `api/v4` and when
referencing merge requests in paths the IID is now needed instead of the
ID.

Tests are passing, but in this case I feel like that doesn't suggest
anything and this branch will undergo some manual testing soon.

[api-docs]: https://docs.gitlab.com/ce/api/v3_to_v4.html