Removed the build number from target url when the state is pending (job is waiting for the next available executor).
Build number in the url causes problems when there are more than one job
in the build queue. In those cases job.getNextBuildNumber returns the same
number for every build in the queue which leads to wrong urls reported to Gitlab.

The build number is assigned when the execution starts so the best option
for pending builds is probably just to set the target url point to the job page without
the build number.