Fix MergeRequestBuildActionTest
Call to MergeRequestBuildAction.execute always thrown an exception and so mock
verifications were never performed - the tests passed because of ExpectedException rule.

All the decisions on whether or not to build a project are inside mergeRequestHookTriggerHandler,
therefore mocking GitLabPushTrigger didn't exercise any of the build triggering logic, and so some
of the tests now fail, demonstrating actual issues.