Actually use the maven-shade-plugin
Previously it was just in the `<build><pluginManagement>` section. It needed
to be in `<build><plugins>`. `<pluginManagement>` lets you define
versions and configuration for plugins that either aren't actually
invoked in your POM, but are invoked in POMs that inherit from your
POM. See, for example,
https://github.com/jenkinsci/plugin-pom/blob/master/pom.xml#L293. In
practice, it also will override `<pluginManagement>` or `<plugins>`
definitions in your POM's parent, but you can do that in `<plugins>`
as well. `<plugins>` defines what will actually be executed in your
POM. So...yeah. Tada?