[JENKINS-62767] Rely on git-client to provide jgit
The git client plugin is intended to be the only plugin in the Jenkins
project that delivers JGit.  Other plugins rely on the version of JGit
inside the git client plugin.  The gitlab-plugin 1.5.13 bundles JGit
3.5.2.  That is a 6 year old version of JGit that has not been used by
the git client plugin for a very long time.  The git client plugin is
currently delivering JGit 5.8.0.

This removes the declared dependency on JGit and relies instead on
the JGit version that is provided by the already declared dependency
on git-client.