Drop ignore certificate errors test
Test relies on a Jetty API that changed from Jetty 9 to Jetty 10.
Test is not worth delaying the use of the most recent parent pom.

Also improves the assertion of the remaining certificate test so that
it is asserting a more precise failure condition.

Confirmed the new assertion passes with Java 11 and with Java 17.