Sync project branch improvements
The repository appears to be initially created as a shallow clone.
We need both branches in order to merge between two branches.  Fetch the
remote repository with the unshallow option.  If that command fails,
then fetch the remote repository without the unshallow option.  The hope
is that one of the two fetch commands will succeed.

Configure the user email address and user name in the repository so
that the merge does not fail with the message that user name and user
email are not configured.  The email address is fake, since no email is
accepted at the jenkins.io site.

Perform the merge with the `--no-edit` command so that merge knows it
should not open an editor.  May be unnecessary, but should not be harmful.