Suppress warnings for spotbugs 4.8.3
https://github.com/jenkinsci/plugin-pom/releases/tag/plugin-4.77
Plugin pom 4.77 needs this change to resolve new spotbugs warnings
that will be reported by spotbugs 4.8.3 and later.

Can skip the changelog, since this is not visible to users.

Confirmed that the spotbugs warnings are visible when using the
4.77-SNAPSHOT plugin pom before this change. With this change, the
spotbugs warnings are no longer visible.

Confirmed that spotbugs is silent with plugin pom 4.77 after this
change.