#1427: Prevent job triggering for non-matching webhooks (#1720)
8ec69c7ada9da38a760f532b9dd34bdf538f8ccb introduced a change that does
not handle jobs created using Jenkins DSL properly, as DSL jobs that do
not specify `triggerOpenMergeRequestOnPush` have the
`triggerOpenMergeRequest` value passed into
`MergeRequestHookTriggerHandlerFactory` as `null`, with the above
change causing the value not to be handled as `never` where the plugin
previously treated `null` as `never`. This results in webhooks
triggering builds for DSL-created jobs even where the job is doing
actions that aren't valid for a Merge Request in that state (e.g.
triggering release builds even though the Merge Request hasn't been
merged). This change returns to handling an unspecified
`triggerOpenMergeRequest` value as being equivalent to specifying it as
`never`.

Co-authored-by: Kris Stern <krisstern@outlook.com>