Fix docker compose port conflict (#1775)
* Fix port conflict

The default jenkins agent port conflicts with the compose file
which currently configures Jenkins with a JAVA_TOOL_OPTIONS
setting that opens JDWP debugging on port 50000.

This commit changes the JDWP debug port to 50001 to resolve the
issues.

Signed-off-by: dodo920306 <dodo920306@gmail.com>

* Remove the Jenkins agent port

It's unnecessary to map the agent port to the host for
development.

Signed-off-by: dodo920306 <dodo920306@gmail.com>

---------

Signed-off-by: dodo920306 <dodo920306@gmail.com>