Check if a connection is unique (#1776)
* Check if a connection is unique

Ensure a connection is unique before adding it to the connection list.

To prevent data inconsistency caused by partially successful updates,
the new list and map are only assigned after validation.

Additionally, GitLabConnectionConfig now overrides
configure(StaplerRequest2, JSONObject) to avoid UI failures with no
clear error messages on it. The validation exception will be caught
and rethrown as FormException.

An error message for duplicate connection names already exists, so
no additional message handling is required.

Signed-off-by: dodo920306 <dodo920306@gmail.com>

* Fix format violations

Signed-off-by: dodo920306 <dodo920306@gmail.com>

* Add unit test

Test if an IllegalArgumentException will be thrown
if there are duplicate connections when
setConnections().

Signed-off-by: dodo920306 <dodo920306@gmail.com>

* Replace statement lambda with expression lambda

Signed-off-by: dodo920306 <dodo920306@gmail.com>

---------

Signed-off-by: dodo920306 <dodo920306@gmail.com>
Co-authored-by: Kris Stern <krisstern@outlook.com>